refactor(engine): Phase<M> trait + consolidate confidence threshold to dedup
Introduce a stateless Phase<M> trait so the orchestrator drives every
per-document step (extraction, detection, deduplication, redaction,
validation) through one uniform interface. Engines + deployment config
move onto RunContext; per-call workflow stays on EngineInput; phases
themselves carry no state beyond a PhantomData modality marker.

Consolidate the confidence-filter knob to one place:
DeduplicationParams.confidence_threshold. Per-recognizer thresholds
(nvisy-pattern PatternEngine, PatternDetection, LlmNerContext,
VlmDetectContext, DetectionContext.score_threshold, Detection workflow,
Redaction workflow override) are gone — per-method skew is folded in via
the calibration map instead.

Other reshapes in the same pass:
- Extractors -> ExtractionEngine; *Section -> *Config; from_config inherent
- ExtractionEngine + DetectionEngine extracted to engine.rs files;
  mod.rs holds only the Phase impl, workflow config, and re-exports
- Validation.fail_on_leak: bool -> OnLeak::{Ignore, Fail} enum
