+++ crates/nvisy-engine/src/deduplication/params.rs
@@ -5,6 +5,7 @@
 //! [`GroupingCriteria`], [`ConflictResolution`]) into the shape the
 //! workflow ingests as a single JSON section.
 
+use nvisy_ontology::primitive::ConfidenceThreshold;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
@@ -17,12 +18,13 @@ use super::resolve::ConflictResolution;
 /// Merges and scores entity candidates from multiple detection
 /// sources into a deduplicated, confidence-scored entity list.
 ///
-/// The minimum confidence threshold isn't on this struct — it comes
-/// from the operator's per-call [`Detection`] config via
-/// [`FilterParams`], so the operator controls filtering uniformly.
+/// Owns the sole confidence threshold in the pipeline: detection
+/// layers and recognizers do not filter on confidence themselves —
+/// per-method skew is folded in via [`calibration`], and the resulting
+/// calibrated score is checked against [`confidence_threshold`] here.
 ///
-/// [`Detection`]: crate::detection::Detection
-/// [`FilterParams`]: super::FilterParams
+/// [`calibration`]: Self::calibration
+/// [`confidence_threshold`]: Self::confidence_threshold
 #[derive(Debug, Clone, Default, PartialEq)]
 #[derive(Serialize, Deserialize, JsonSchema)]
 pub struct DeduplicationParams {
@@ -32,9 +34,13 @@ pub struct DeduplicationParams {
     /// Strategy for combining confidence scores.
     #[serde(default)]
     pub strategy: DeduplicationStrategy,
-    /// Per-method confidence scaling applied before deduplication.
+    /// Per-method confidence scaling applied before filtering.
     #[serde(default, skip_serializing_if = "std::collections::HashMap::is_empty")]
     pub calibration: CalibrationMap,
+    /// Minimum (calibrated) confidence an entity must clear to survive
+    /// deduplication. `None` keeps every candidate.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub confidence_threshold: Option<ConfidenceThreshold>,
     /// How to resolve conflicts when different entity kinds overlap
     /// the same text span.
     #[serde(default)]
