+++ crates/nvisy-engine/src/validation/workflow.rs
@@ -8,6 +8,20 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
+/// How the validation phase reacts when post-redaction re-scan finds
+/// a value that should have been redacted but still appears in the
+/// output.
+#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
+#[derive(Serialize, Deserialize, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub enum OnLeak {
+    /// Log the leak and continue. The run still succeeds.
+    #[default]
+    Ignore,
+    /// Fail the run with a validation error listing the leaked values.
+    Fail,
+}
+
 /// Controls how the post-redaction leak check affects the overall pipeline
 /// outcome.
 ///
@@ -15,7 +29,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
 #[derive(Serialize, Deserialize, JsonSchema)]
 pub struct Validation {
-    /// Fail the run if any leaked values are detected.
+    /// What to do when the re-scan finds a leaked value.
     #[serde(default)]
-    pub fail_on_leak: bool,
+    pub on_leak: OnLeak,
 }
